requirements.txt
anyio==4.12.1
certifi==2026.1.4
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-dotenv==1.2.1
python-telegram-bot==22.5
typing_extensions==4.15.0
